Replace test package with testv2 package
