+++ Makefile
@@ -72,7 +72,7 @@ buildenv:
 .PHONY: lint
 lint:
 	@golangci-lint version
-	golangci-lint run --timeout=5m -v ./pkg/... ./test/... ./testv2/...
+	golangci-lint run --timeout=5m -v ./pkg/... ./test/...
 
 .PHONY: verify-licence
 verify-licence: GOFLAGS = -mod=readonly
@@ -105,4 +105,4 @@ fmt: shfmt prowfmt
 
 gogenerate:
 	go generate ./pkg/...
-	go generate ./testv2/...
+	go generate ./test/...
