--- test/e2e/flag_test.go
@@ -1,58 +1,0 @@
-/*
-Copyright 2019 The KubeOne Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package e2e
-
-import (
-	"testing"
-)
-
-func Test_containerRuntimeFlag_Set(t *testing.T) {
-	tests := []struct {
-		name    string
-		text    string
-		wantErr bool
-	}{
-		{
-			name: "empty",
-			text: "",
-		},
-		{
-			name: "docker",
-			text: "docker",
-		},
-		{
-			name: "containerd",
-			text: "containerd",
-		},
-		{
-			name:    "gibberish",
-			text:    "gibberish",
-			wantErr: true,
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			crf := &containerRuntimeFlag{}
-
-			if err := crf.Set(tt.text); (err != nil) != tt.wantErr {
-				t.Errorf("containerRuntimeFlag.Set() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
