+++ test/e2e/scenario_install.go
