+++ test/e2e/sonobuoy.go
@@ -24,7 +24,7 @@ import (
 	"io"
 	"os"
 
-	"k8c.io/kubeone/test/e2e/testutil"
+	"k8c.io/kubeone/test/exec"
 )
 
 const sonobuoyResultsFile = "results.tar.gz"
@@ -111,16 +111,16 @@ func (sbb *sonobuoyBin) run(ctx context.Context, args ...string) error {
 	return sbb.build(args...).BuildCmd(ctx).Run()
 }
 
-func (sbb *sonobuoyBin) build(args ...string) *testutil.Exec {
-	exe := testutil.NewExec("sonobuoy",
-		testutil.WithArgs(args...),
-		testutil.WithEnv(os.Environ()),
-		testutil.InDir(sbb.dir),
-		testutil.StdoutDebug,
+func (sbb *sonobuoyBin) build(args ...string) *exec.Exec {
+	exe := exec.NewExec("sonobuoy",
+		exec.WithArgs(args...),
+		exec.WithEnv(os.Environ()),
+		exec.InDir(sbb.dir),
+		exec.StdoutDebug,
 	)
 
 	if sbb.kubeconfig != "" {
-		testutil.WithEnvs(fmt.Sprintf("KUBECONFIG=%s", sbb.kubeconfig))(exe)
+		exec.WithEnvs(fmt.Sprintf("KUBECONFIG=%s", sbb.kubeconfig))(exe)
 	}
 
 	if sbb.proxyURL != "" {
@@ -128,7 +128,7 @@ func (sbb *sonobuoyBin) build(args ...string) *testutil.Exec {
 			fmt.Sprintf("HTTPS_PROXY=%s", sbb.proxyURL),
 			fmt.Sprintf("HTTP_PROXY=%s", sbb.proxyURL),
 		}
-		testutil.WithEnv(proxies)(exe)
+		exec.WithEnv(proxies)(exe)
 	}
 
 	return exe
