+++ test/e2e/testdata/containerd_cilium.yaml
