+++ test/e2e/testdata/containerd_simple_external.yaml
