+++ test/e2e/testdata/docker_calico.yaml
