+++ test/e2e/testdata/docker_cilium.yaml
