+++ test/e2e/testdata/docker_simple.yaml
