+++ test/e2e/testdata/docker_simple_external.yaml
