+++ test/e2e/testdata/docker_weave.yaml
