+++ test/e2e/testdata/kube_proxy_ipvs.yaml
