+++ test/e2e/testdata/legacy_machine_controller_containerd.yaml
