+++ test/e2e/testdata/legacy_machine_controller_containerd_external.yaml
