+++ test/e2e/testdata/legacy_machine_controller_docker_external.yaml
