+++ test/e2e/testdata/openstack_centos.tfvars
