+++ test/e2e/testdata/openstack_rockylinux.tfvars
